Se coloca el humanize en humanize.js y se exporta la función
